+++ NEWS.md
@@ -5,6 +5,7 @@
 ## New functions
 
 * `ir_scale()`: Function to scale intensity values in spectra, similar to `base::scale()`. Intensity values are scaled for the same x axis values.
+* `ir_as_ir.hyperSpec()`: Conversion of `hyperSpec::hyperSpec` objects to `ir` objects. 
 
 ## Improvements